% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconstructSigs.R
\docType{package}
\name{deconstructSigs}
\alias{deconstructSigs-package}
\alias{deconstructSigs}
\title{deconstructSigs}
\description{
Takes sample information in the form of the fraction of mutations
in each of 96 trinucleotide contexts and identifies the weighted combination 
of published signatures that, when summed, most closely reconstructs the 
mutational profile.
}
\section{Main functions}{

\itemize{ 
  \item \code{\link{whichSignatures}} 
  \item \code{\link{mut.to.sigs.input}} 
  \item \code{\link{plotSignatures}} 
  \item \code{\link{makePie}}
  }
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/raerose01/deconstructSigs}
  \item Report bugs at \url{https://github.com/raerose01/deconstructSigs/issues}
}

}
\author{
Rachel Rosenthal rachel.rosenthal.14@ucl.ac.uk
}
