% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{makePie}
\alias{makePie}
\title{Plots the weights from whichSignatures()}
\usage{
makePie(sigs.output, sub = "", v3 = FALSE, add.color = NULL)
}
\arguments{
\item{sigs.output}{The list output from whichSignatures()}

\item{sub}{A character vector that specifies cancer subtype for plot title, 
if wanted}

\item{add.color}{Optional character vector to assign additional colors for
novel signatures}
}
\value{
Plots a pie chart of the weights calculated in the given tumor sample
}
\description{
Uses the output from whichSignatures() and creates a pie chart of the weights
outputted
}
\examples{
makePie(example.output)
}
