% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{signatures.exome.cosmic.v3.may2019}
\alias{signatures.exome.cosmic.v3.may2019}
\title{Updated SBS Exome Signatures from Sanger COSMIC (May 2019)}
\format{
A data frame of 65 rows and 96 columns
}
\source{
\url{https://www.synapse.org/#!Synapse:syn12009743}
}
\description{
A dataset containing the additional signatures identified read into R as a 
data frame. Can be used the 'signatures.ref' parameter in whichSignatures().
}
\keyword{datasets}
